inor fixing and refactor to HelloControllerTest
