+++ src/main/java/com/labforward/api/core/exception/EntityValidationException.java
@@ -2,14 +2,14 @@ package com.labforward.api.core.exception;
 
 import org.springframework.validation.BindingResult;
 
-import static com.labforward.api.constants.Messages.BAD_REQUEST;
+import static com.labforward.api.constants.Messages.UNPROCESSEABLE_ENTIIIY;
 
 public class EntityValidationException extends RuntimeException {
 
     private final transient BindingResult bindingResult;
 
     public EntityValidationException(BindingResult bindingResult) {
-        super(BAD_REQUEST);
+        super(UNPROCESSEABLE_ENTIIIY);
         this.bindingResult = bindingResult;
     }
