+++ src/main/java/com/labforward/api/core/validation/EntityValidator.java
@@ -11,49 +11,50 @@ import static com.labforward.api.constants.Messages.MESSAGE_NO_ID_MATCH;
 
 public class EntityValidator {
 
-	private SpringValidatorAdapter validatorAdapter;
+    private SpringValidatorAdapter validatorAdapter;
 
-	@Autowired
-	public EntityValidator(SpringValidatorAdapter validatorAdapter) {
-		this.validatorAdapter = validatorAdapter;
-	}
+    @Autowired
+    public EntityValidator(SpringValidatorAdapter validatorAdapter) {
+        this.validatorAdapter = validatorAdapter;
+    }
 
-	public void validateCreate(Object target, Object... groups) throws EntityValidationException {
-		validate(target, groups);
-	}
+    public void validateCreate(Object target, Object... groups) throws EntityValidationException {
+        validate(target, groups);
+    }
 
-	public void validateUpdate(Object target, Object... groups) throws EntityValidationException {
-		validate(target, groups);
-	}
+    public void validateUpdate(Object target, Object... groups) throws EntityValidationException {
+        validate(target, groups);
+    }
 
-	public void validateUpdate(String id, Entity target) throws EntityValidationException {
-		Preconditions.checkArgument(id != null);
-		Preconditions.checkArgument(target != null);
+    public void validateUpdate(String id, Entity target) throws EntityValidationException {
+        Preconditions.checkArgument(id != null);
+        Preconditions.checkArgument(target != null);
 
-		BeanPropertyBindingResult result = new BeanPropertyBindingResult(target, target.getClass().getName());
+        BeanPropertyBindingResult result = new BeanPropertyBindingResult(target, target.getClass().getName());
 
-		// validate ids match
-		if (!id.equals(target.getId())) {
-			FieldError fieldError = new FieldError(target.getClass().getName(), "id", MESSAGE_NO_ID_MATCH);
-			result.addError(fieldError);
-			throw new EntityValidationException(result);
-		}
+        // validate ids match
+        if (!id.equals(target.getId())) {
+            FieldError fieldError = new FieldError(target.getClass().getName(), "id", MESSAGE_NO_ID_MATCH);
+            result.addError(fieldError);
+            throw new EntityValidationException(result);
+        }
 
-		validatorAdapter.validate(target, result, EntityUpdateValidatorGroup.class);
+        validatorAdapter.validate(target, result, EntityUpdateValidatorGroup.class);
 
-		if (result.hasErrors()) {
-			throw new EntityValidationException(result);
-		}
-	}
+        if (result.hasErrors()) {
+            throw new EntityValidationException(result);
+        }
+    }
 
-	protected void validate(Object target, Object... groups) throws EntityValidationException {
-		Preconditions.checkArgument(target != null);
+    protected void validate(Object target, Object... groups) throws EntityValidationException {
+        Preconditions.checkNotNull(target);
 
-		BeanPropertyBindingResult result = new BeanPropertyBindingResult(target, target.getClass().getName());
-		validatorAdapter.validate(target, result, groups);
 
-		if (result.hasErrors()) {
-			throw new EntityValidationException(result);
-		}
-	}
+        BeanPropertyBindingResult result = new BeanPropertyBindingResult(target, target.getClass().getName());
+        validatorAdapter.validate(target, result, groups);
+
+        if (result.hasErrors()) {
+            throw new EntityValidationException(result);
+        }
+    }
 }
