+++ src/main/java/com/labforward/api/hello/controller/HelloController.java
@@ -11,6 +11,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.validation.Valid;
+
 import static com.labforward.api.constants.Messages.DEFAULT_ID;
 
 @RestController
